# Avoid OOM in `normaliseIterationData`