--- a/lib/runner/extensions/waterfall.command.js
+++ b/lib/runner/extensions/waterfall.command.js
@@ -3,7 +3,8 @@ var _ = require('lodash'),
     VariableScope = require('postman-collection').VariableScope,
 
     prepareLookupHash,
-    extractSNR;
+    extractSNR,
+    getIterationData;
 
 /**
  * Returns a hash of IDs and Names of items in an array
@@ -39,6 +40,21 @@ extractSNR = function (executions, previous) {
     });
 
     return snr;
+};
+
+/**
+ * Returns the data for the given iteration
+ *
+ * @function getIterationData
+ * @param {Array} data - The data array containing all iterations' data
+ * @param {Number} iteration - The iteration to get data for
+ * @return {Any} - The data for the iteration
+ */
+getIterationData = function (data, iteration) {
+    // check if data exists. if not, return the last member from the data
+    if (data && data.length) {
+        return iteration in data ? data[iteration] : data[data.length - 1];
+    }
 };
 
 /**
@@ -139,9 +155,7 @@ module.exports = {
                 this.queue('item', {
                     item: item,
                     coords: coords,
-                    data: this.state.data[coords.iteration] === undefined ?
-                        this.state.data[this.state.data.length - 1] :
-                        this.state.data[coords.iteration],
+                    data: getIterationData(this.state.data, coords.iteration),
                     environment: this.state.environment,
                     globals: this.state.globals,
                     collectionVariables: this.state.collectionVariables,