--- a/lib/runner/extensions/waterfall.command.js
+++ b/lib/runner/extensions/waterfall.command.js
@@ -139,7 +139,9 @@ module.exports = {
                 this.queue('item', {
                     item: item,
                     coords: coords,
-                    data: this.state.data[coords.iteration],
+                    data: this.state.data[coords.iteration] === undefined ?
+                        this.state.data[this.state.data.length - 1] :
+                        this.state.data[coords.iteration],
                     environment: this.state.environment,
                     globals: this.state.globals,
                     collectionVariables: this.state.collectionVariables,