--- a/lib/runner/extensions/waterfall.command.js
+++ b/lib/runner/extensions/waterfall.command.js
@@ -51,10 +51,13 @@ extractSNR = function (executions, previous) {
  * @return {Any} - The data for the iteration
  */
 getIterationData = function (data, iteration) {
-    // check if data exists. if not, return the last member from the data
-    if (data && data.length) {
-        return iteration in data ? data[iteration] : data[data.length - 1];
+    // if iteration has a corresponding data element use that
+    if (iteration < data.length) {
+        return data[iteration];
     }
+
+    // otherwise use the last data element
+    return data[data.length - 1];
 };
 
 /**
@@ -84,7 +87,7 @@ module.exports = {
         // if the location in state is already normalised then go ahead and queue iteration, else normalise the
         // location
         state.cursor = Cursor.box(state.cursor, { // we pass bounds to ensure there is no stale state
-            cycles: state.data.length,
+            cycles: this.options.iterationCount,
             length: state.items.length
         });
         this.waterfall = state.cursor; // copy the location object to instance for quick access