--- a/lib/runner/index.js
+++ b/lib/runner/index.js
@@ -135,12 +135,12 @@ _.assign(Runner, {
         !_.isArray(data) && (data = [{}]);
         ((length < 1) || !_.isFinite(length)) && (length = data.length);
 
-        // if the iteration count is greater than the length of data, we repeat the last item in data. here, we are
-        // using a for-loop so that we do not have excessive spatial complexity
+        // if the iteration count is greater than the length of data, we
+        // increase the length of data to match the number of iterations.
+        // We set the last element of data as the one we will use for all
+        // elements of data which are undefined.
         if (length > data.length) {
-            for (var i = data.length, filler = data[i - 1]; i < length; i++) {
-                data[i] = filler;
-            }
+            data[length - 1] = data[data.length - 1];
         }
 
         // just to be sure that there are no extra items in array, we match the lengths and return the data