--- a/lib/runner/index.js
+++ b/lib/runner/index.js
@@ -102,9 +102,17 @@ _.assign(Runner.prototype, {
                 return callback(new Error(`Unable to find a folder or request: ${_.get(options, 'entrypoint.execute')}`));
             }
 
+            // ensure data is an array
+            !_.isArray(options.data) && (options.data = [{}]);
+
+            // get iterationCount from data if not set
+            if (!runOptions.iterationCount) {
+                runOptions.iterationCount = options.data.length;
+            }
+
             return callback(null, (new Run({
                 items: runnableItems,
-                data: Runner.normaliseIterationData(options.data, options.iterationCount),
+                data: options.data,
                 environment: options.environment,
                 globals: _.has(options, 'globals') ? options.globals : self.options.globals,
                 // @todo Move to item level to support Item and ItemGroup variables
@@ -122,32 +130,7 @@ _.assign(Runner, {
      *
      * @type {Run}
      */
-    Run: Run,
-
-    /**
-     * @private
-     *
-     * @param {Array} data
-     * @param {Number} length
-     * @returns {Array}
-     */
-    normaliseIterationData: function (data, length) {
-        !_.isArray(data) && (data = [{}]);
-        ((length < 1) || !_.isFinite(length)) && (length = data.length);
-
-        // if the iteration count is greater than the length of data, we
-        // increase the length of data to match the number of iterations.
-        // We set the last element of data as the one we will use for all
-        // elements of data which are undefined.
-        if (length > data.length) {
-            data[length - 1] = data[data.length - 1];
-        }
-
-        // just to be sure that there are no extra items in array, we match the lengths and return the data
-        data.length = length;
-
-        return data;
-    }
+    Run: Run
 });
 
 module.exports = Runner;