--- a/test/unit/runner.test.js
+++ b/test/unit/runner.test.js
@@ -2,6 +2,10 @@ var expect = require('chai').expect,
     Runner = require('../../lib/runner/'),
     sdk = require('postman-collection');
 
+function getIterationData (data, iteration) {
+    return iteration in data ? data[iteration] : data[data.length - 1];
+}
+
 describe('runner', function () {
     describe('Run', function () {
         it('should be a constructor', function () {
@@ -328,7 +332,14 @@ describe('runner', function () {
         });
 
         it('should duplicate the last element of the data set if length is greater', function () {
-            expect(Runner.normaliseIterationData([{foo: 'alpha'}], 2)).to.eql([{foo: 'alpha'}, {foo: 'alpha'}]);
+            var data = Runner.normaliseIterationData([{foo: 'alpha'}, undefined, {bar: 'beta'}], 6);
+
+            expect(getIterationData(data, 0)).to.eql({foo: 'alpha'});
+            expect(getIterationData(data, 1)).to.eql(undefined);
+            expect(getIterationData(data, 2)).to.eql({bar: 'beta'});
+            expect(getIterationData(data, 3)).to.eql({bar: 'beta'});
+            expect(getIterationData(data, 4)).to.eql({bar: 'beta'});
+            expect(getIterationData(data, 5)).to.eql({bar: 'beta'});
         });
     });
 });