### Description

If iteration count is more than the length of `data`, the `normaliseIterationData` function fills the `data` array with the last element in `data` such that `data.length = iterationCount`.

When iteration count is a large number (e.g. 99999999), run creation takes a lot of time and sometimes even fails with "Out of Memory" error because the JS Heap is exhausted.

This change avoids the above situation by simply setting the `data[iterationCount - 1]` element as the value to be repeated. Later, when iteration data is to be fetched, if the element at `data[iteration]` exists, we use it; if not, we use the last element of the `data`.

---

### Performance increase
For `iterationCount` 9999999, time for `run` creation:
#### before
- 3s 417ms
- 3s 403ms
- 3s 505ms

#### after
- 0s 2.1ms
- 0s 1.9ms
- 0s 1.9ms